refactor(disputes): extract transfer-recovery core for reuse
Move the charge->transfer->reversal math (metadata dispute_id replay
guard, min(share, unreversed) cap, dispute-reversal-{id} idempotency key)
from registrations.ts into dispute-recovery.ts recoverDisputedTransfer.
The caller passes an auditNote sink (namespaces + injects entity/org
context) and reversalMetadata (entity id on the reversal). registrations
keeps a thin wrapper; behaviour bit-identical (dispute suites 54/54).
Task 6 (sponsors) / Task 7 (platform) consume the same core.
